Mortgage assistant: Avoid crash when returning to the repayment page.
And as a side benefit don't leak signal handlers for the repGncFreq
changed signal.

The crash occured because gnc_frequency_setup_recurrence first calls
gnc_date_edit_set_gdate which emits a changed signal that, if it
wasn't the first call to loan_rep_prep would call
update_repayment_formula_cb and that would generate a new recurrences
list. But the still running gnc_frequency_setup_recurrence still had a
pointer to the old list and when it tried to manipulate it it would
crash.

This change disconnects the previous signal handler preventing the
list from being replaced.