[report-core.scm] add report anchor mechanism
The <report> object contains the data for a report with options. Add a
mechanism to store the list of anchors contained in it; adding an
anchor will return the unique integer key referring to the report's
anchor; retrieval will return the anchor.

Note: each anchor will typically be: (list reportname src-options
optionlist)

whereby reportname is the desired report guid, src-options is the
originating report options object, and optionlist is a list of options to be
set.