[gnc-locale-name] Use setlocale(LC_MONETARY, NULL) to retrieve name.
Instead of setlocale(LC_ALL, NULL). The uses for the string either create
an ICU locale or sort between German and US TXF reports. On some OSes
where the user has different locales set for one or more categories LC_ALL
produces a dump of all of the LC_FOO variables and their values. That's
not useful for comparing to de_DE and it's too long for ICU, which
crashes.

Meanwhile in dialog-print-checks the RV of setlocale was ignored,
rendering the call a no-op.